add consolidation & withdrawal matchers

+++ db/schema/sqlite/20240805095505_pectra-updates2.sql
@@ -4,7 +4,7 @@
 DROP TABLE IF EXISTS "consolidations";
 
 CREATE TABLE IF NOT EXISTS "consolidation_requests" (
-    slot_number INT NOT NULL,
+    slot_number BIGINT NOT NULL,
     slot_root BLOB NOT NULL,
     slot_index INT NOT NULL,
     orphaned bool NOT NULL DEFAULT FALSE,
@@ -42,7 +42,7 @@ CREATE INDEX IF NOT EXISTS "consolidation_requests_fork_idx"
 DROP TABLE IF EXISTS "el_requests";
 
 CREATE TABLE IF NOT EXISTS "withdrawal_requests" (
-    slot_number INT NOT NULL,
+    slot_number BIGINT NOT NULL,
     slot_root BLOB NOT NULL,
     slot_index INT NOT NULL,
     orphaned bool NOT NULL DEFAULT FALSE,
